WT-3076 Add a general-purpose epoch manager (#3362)
* Switch the table schema generation numbers over the general purpose generations manager.

* Switch the btree split generation numbers over the general purpose generations manager.

* Add replaced hazard pointer arrays to the session's stash for later free (__wt_txn_release() and all functions that release session resources will free old hazard pointer arrays where possible).

* The connection's free-on-close functionality is no longer used outside of the config code, make it private again.

* To convert hazard pointers to the standard generation code, we have to add enter/leave segments when looking at other session's hazard pointer arrays. (The only paths freeing hazard
pointer memory are in the context of the thread that "owns" the memory, so there's no need to enter the hazard pointer resource generation when the thread that owns the hazard pointer memory is examining it, it can't be freed from underneath that thread.)

+++ src/include/schema.h
@@ -68,8 +68,8 @@ struct __wt_table {
 	bool cg_complete, idx_complete, is_simple;
 	u_int ncolgroups, nindices, nkey_columns;
 
-	uint32_t refcnt;	/* Number of open cursors */
-	uint32_t schema_gen;	/* Cached schema generation number */
+	uint32_t refcnt;		/* Number of open cursors */
+	uint64_t schema_gen;		/* Cached schema generation number */
 };
 
 /*
